+++ fatfs/src/versions/exfat.rs
@@ -7,8 +7,8 @@ pub struct ExFatBpb {
     pub jmp_boot: [u8; 3],
     pub oem_name: [u8; 8],
     pub padding: [u8; 53],
-    pub partition_offset: u64,
-    pub vol_length: u64,
+    pub partition_offset: usize,
+    pub vol_length: usize,
     pub fat_offset: u32,
     pub fat_length: u32,
     pub cluster_heap_offset: u32,
@@ -28,23 +28,23 @@ pub struct ExFatBpb {
 pub struct ExFatOps {
     pub bytes_per_sector: u32,
     pub sectors_per_cluster: u32,
-    pub fat_start_sector: u64,
-    pub data_start_sector: u64,
+    pub fat_start_sector: usize,
+    pub data_start_sector: usize,
     pub root_cluster: u32,
 }
 
 impl FatOps for ExFatOps {
     fn get_next_cluster(&self, reader: &BlockReader, cluster: u32) -> Result<u32, Error> {
         // exFAT FAT entries are 32-bit
-        let fat_offset = cluster as u64 * 4;
-        let fat_sector_offset = fat_offset / self.bytes_per_sector as u64;
-        let entry_offset = (fat_offset % self.bytes_per_sector as u64) as usize;
+        let fat_offset = cluster as usize * 4;
+        let fat_sector_offset = fat_offset / self.bytes_per_sector as usize;
+        let entry_offset = (fat_offset % self.bytes_per_sector as usize) as usize;
 
         let sector = self.fat_start_sector + fat_sector_offset;
 
         // TODO: Handle buffer size dynamically if sector > 512
         let mut buf = alloc::vec![0u8; self.bytes_per_sector as usize];
-        let read_pos = sector * self.bytes_per_sector as u64;
+        let read_pos = sector * self.bytes_per_sector as usize;
         reader.read_offset(read_pos, &mut buf).map_err(|_| Error::IoError)?;
 
         let ptr = unsafe { buf.as_ptr().add(entry_offset) };
@@ -53,10 +53,10 @@ impl FatOps for ExFatOps {
         Ok(val) // All 32 bits are valid
     }
 
-    fn cluster_to_sector(&self, cluster: u32) -> u64 {
+    fn cluster_to_sector(&self, cluster: u32) -> usize {
         // exFAT 1st cluster is cluster 2 usually
         let rel_cluster = if cluster >= 2 { cluster - 2 } else { 0 };
-        self.data_start_sector + (rel_cluster as u64 * self.sectors_per_cluster as u64)
+        self.data_start_sector + (rel_cluster as usize * self.sectors_per_cluster as usize)
     }
 
     fn get_root_location(&self) -> RootLocation {
